move RandomLinuxIsos into utils/index
